Refactor chart rendering and enhance query result handling
- Update Chart.js integration by implementing an idempotent registration function to streamline chart rendering.
- Introduce new components for rendering charts and handling query results, including lazy-loaded tabs for charts and explanations.
- Add utility functions for managing edit diffs and rendering changes in query results.
- Enhance the overall structure of the renderer by separating concerns and improving modularity.

These changes improve the user experience in visualizing data and managing query results within the application.
